# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.QueryRequest:
    properties:
      filter:
        additionalProperties: true
        description: 'Example: {"address.city": "London", "age": "gt:18"}'
        type: object
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      sort:
        example: address.city:desc
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A multi-tenant JSONB document store for students.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VelocityDBGo API
  version: "1.0"
paths:
...
  /api/projects/{projectId}/data/{collectionName}:
    get:
      description: Retrieve all documents from a collection. Supports dynamic filtering
        via query parameters (e.g. ?name=John, ?address.city=London, ?age=gte:18),
        sorting (e.g. ?sort=address.zip:desc), and pagination (?limit=10&offset=0).
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection Name
        in: path
        name: collectionName
        required: true
        type: string
      - description: Pagination limit
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Sort field and direction (e.g., age:desc, address.city:asc)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get all documents
      tags:
      - Data
    post:
      consumes:
      - application/json
      description: Insert an arbitrary JSON payload into a collection
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection Name
        in: path
        name: collectionName
        required: true
        type: string
      - description: JSON Data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Insert JSON document
      tags:
      - Data
  /api/projects/{projectId}/data/{collectionName}/{docId}:
    delete:
      description: Delete a document from a collection by ID
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection Name
        in: path
        name: collectionName
        required: true
        type: string
      - description: Document ID
        in: path
        name: docId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete a document
      tags:
      - Data
    get:
      description: Retrieve a specific document from a collection by ID
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection Name
        in: path
        name: collectionName
        required: true
        type: string
      - description: Document ID
        in: path
        name: docId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get a single document
      tags:
      - Data
    put:
      consumes:
      - application/json
      description: Fully replace or update the JSON payload of an existing document
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection Name
        in: path
        name: collectionName
        required: true
        type: string
      - description: Document ID
        in: path
        name: docId
        required: true
        type: integer
      - description: Updated JSON Data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update a document
      tags:
      - Data
  /api/projects/{projectId}/data/{collectionName}/query:
    post:
      consumes:
      - application/json
      description: Query documents using a structured JSON POST payload. Useful when
        URL query parameters become too complex.
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection Name
        in: path
        name: collectionName
        required: true
        type: string
      - description: Query Parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.QueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Query documents via JSON
      tags:
      - Data
...
securityDefinitions:
  ApiKeyAuth:
    description: Public API Key for a specific Project.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"